// web/src/lib/api/generated/model/validationErrorResponse.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Spark API
 * Spark API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationErrorResponseErrorsItem } from './validationErrorResponseErrorsItem';

export interface ValidationErrorResponse {
  code: string;
  message: string;
  errors: ValidationErrorResponseErrorsItem[];
}